fix custom lib import
